refactor: 팀 선택 페이지 스텝 컴포넌트 네이밍 정리 #249
- ReferenceData.tsx → Step3ReferenceData.tsx
- Step3Timeline.tsx → Step4Timeline.tsx
- Step4TeamSelect.tsx → Step5TeamSelect.tsx
- 테스트 파일도 동일하게 변경
- index.tsx import 경로 업데이트
