+++ frontend/src/pages/teamSelectPage/components/steps/Step5TeamSelect.tsx
